# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Astral Core v7 API
  version: 7.0.0
  description: |
    Comprehensive mental health platform API providing secure endpoints for:
    - User authentication and MFA management
    - Crisis assessment and intervention
    - Wellness data tracking and analytics
    - Payment processing and subscription management
    - Appointment scheduling and session management
    - Secure messaging and file management
    
    **Security Features:**
    - Rate limiting on all endpoints
    - PHI-compliant data encryption
    - Comprehensive audit logging
    - Role-based access control
    
    **Base URL:** `https://astral-core.app/api`
    
  contact:
    name: Astral Core API Support
    email: api-support@astral-core.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://astral-core.app/api
    description: Production server
  - url: https://staging.astral-core.app/api
    description: Staging server
  - url: http://localhost:3000/api
    description: Local development server
security:
  - bearerAuth: []
  - sessionAuth: []
paths:
...
  /payments/subscriptions:
    get:
      tags: [Payments, Subscriptions]
      summary: Get user subscription
      description: Retrieve current user's subscription details
      operationId: getUserSubscription
      responses:
        '200':
          description: Subscription details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags: [Payments, Subscriptions]
      summary: Create subscription
      description: |
        Create a new subscription for a therapy plan.
        Automatically handles payment method setup if needed.
      operationId: createSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
            examples:
              with_payment_method:
                summary: With existing payment method
                value:
                  therapyPlanId: "plan_therapy_basic"
                  paymentMethodId: "pm_1234567890"
                  couponCode: "SAVE20"
              setup_required:
                summary: Requires payment setup
                value:
                  therapyPlanId: "plan_therapy_premium"
      responses:
        '201':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSubscriptionResponse'
              examples:
                immediate_active:
                  summary: Subscription immediately active
                  value:
                    subscription:
                      id: "sub_1234567890"
                      status: "ACTIVE"
                      planName: "Basic Therapy Plan"
                      amount: 99.00
                    success: true
                    message: "Subscription created successfully"
                setup_required:
                  summary: Payment setup required
                  value:
                    subscription:
                      id: "sub_0987654321"
                      status: "INCOMPLETE"
                      planName: "Premium Therapy Plan"
                      amount: 199.00
                    setupIntent:
                      id: "seti_1234567890"
                      clientSecret: "seti_1234567890_secret_abc123"
                    success: true
                    message: "Subscription created. Please complete payment setup."
        '400':
          description: Invalid request or user already has subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/ServerError'

    put:
      tags: [Payments, Subscriptions]
      summary: Update subscription
      description: Change subscription plan with proration options
      operationId: updateSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubscriptionRequest'
      responses:
        '200':
          description: Subscription updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      tags: [Payments, Subscriptions]
      summary: Cancel subscription
      description: Cancel user subscription with options for immediate or period-end cancellation
      operationId: cancelSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelSubscriptionRequest'
      responses:
        '200':
          description: Subscription canceled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
  schemas:
...
    CreateSubscriptionRequest:
      type: object
      required:
        - therapyPlanId
      properties:
        therapyPlanId:
          type: string
          description: ID of the therapy plan to subscribe to
          example: "plan_therapy_basic"
        paymentMethodId:
          type: string
          description: ID of existing payment method (optional)
          example: "pm_1234567890"
        couponCode:
          type: string
          description: Optional coupon code for discount
          example: "SAVE20"
    UpdateSubscriptionRequest:
      type: object
      required:
        - newTherapyPlanId
      properties:
        newTherapyPlanId:
          type: string
          description: ID of the new therapy plan
        prorationBehavior:
          type: string
          enum: [create_prorations, none, always_invoice]
          description: How to handle proration
          default: create_prorations
    CancelSubscriptionRequest:
      type: object
      required:
        - subscriptionId
      properties:
        subscriptionId:
          type: string
          description: ID of subscription to cancel
        cancelAtPeriodEnd:
          type: boolean
          description: Cancel at period end (true) or immediately (false)
          default: true
        cancellationReason:
          type: string
          description: Optional reason for cancellation
          example: "No longer needed"
...
    SubscriptionResponse:
      type: object
      properties:
        subscription:
          type: object
          properties:
            id:
              type: string
            status:
              type: string
              enum: [INCOMPLETE, INCOMPLETE_EXPIRED, TRIALING, ACTIVE, PAST_DUE, CANCELED, UNPAID, PAUSED]
            planName:
              type: string
            amount:
              type: number
            currency:
              type: string
            interval:
              type: string
            currentPeriodStart:
              type: string
              format: date-time
            currentPeriodEnd:
              type: string
              format: date-time
            cancelAt:
              type: string
              format: date-time
              nullable: true
            trialEnd:
              type: string
              format: date-time
              nullable: true
        success:
          type: boolean
    CreateSubscriptionResponse:
      type: object
      properties:
        subscription:
          type: object
          properties:
            id:
              type: string
            status:
              type: string
            planName:
              type: string
            amount:
              type: number
        setupIntent:
          type: object
          properties:
            id:
              type: string
            clientSecret:
              type: string
          description: Setup intent for payment method setup (if required)
        success:
          type: boolean
        message:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "An error occurred"
        details:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
              path:
                type: array
                items:
                  type: string
          description: Detailed validation errors (when applicable)

  # ==================== COMMON RESPONSES ====================
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              summary: Unauthorized access
              value:
                error: "Authentication required"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation_failed:
              summary: Validation failed
              value:
                error: "Validation failed"
                details:
                  - code: "invalid_type"
                    message: "Expected string, received number"
                    path: ["email"]
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            rate_limited:
              summary: Too many requests
              value:
                error: "Too many requests. Please try again later."
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            server_error:
              summary: Server error
              value:
                error: "An internal server error occurred"
...